requirements.txt
Flask==2.1.1
SQLAlchemy==1.4.22
Jinja2==3.0.2
boto3
mysql-connector-python==8.0.28
Werkzeug==2.2.2
python-dotenv

